--- Fase 2/Evidencias Proyecto/proyecto/app ionic/Comandos Andrea.txt
@@ -1,1 +1,0 @@
-ng generate interface interfaces/usuario
