+++ Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/.vscode/settings.json
@@ -1,3 +1,6 @@
 {
-  "typescript.preferences.autoImportFileExcludePatterns": ["@ionic/angular/common", "@ionic/angular"]
+  "typescript.preferences.autoImportFileExcludePatterns": [
+    "@ionic/angular/common",
+    "@ionic/angular"
+  ]
 }
