--- Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/src/app/interfaces/familiar-registrado.ts
@@ -1,14 +1,0 @@
-export interface FamiliarRegistrado {
-    id_usuario: number;
-    telefono: string;
-  }
-  
-  export interface ContactoAPI {
-    nombre: string;
-    telefono: string;
-  }
-  
-  export interface RespuestaContactosAPI extends FamiliarRegistrado {
-    contactos: ContactoAPI[];
-  }
-
