--- Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/src/app/pages/obtener-contactos/obtener-contactos.page.spec.ts
@@ -1,17 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { ObtenerContactosPage } from './obtener-contactos.page';
-
-describe('ObtenerContactosPage', () => {
-  let component: ObtenerContactosPage;
-  let fixture: ComponentFixture<ObtenerContactosPage>;
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(ObtenerContactosPage);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
