--- Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/src/app/services/api-obtener-contactos.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ApiObtenerContactosService } from './api-obtener-contactos.service';
-
-describe('ApiObtenerContactosService', () => {
-  let service: ApiObtenerContactosService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ApiObtenerContactosService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
