--- Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/src/app/services/api-prueba.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ApiPruebaService } from './api-prueba.service';
-
-describe('ApiPruebaService', () => {
-  let service: ApiPruebaService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ApiPruebaService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
