--- Fase 2/Evidencias Proyecto/proyecto/app ionic/TatasApp/src/app/services/api-usuarios.service.spec.ts
@@ -1,16 +1,0 @@
-import { TestBed } from '@angular/core/testing';
-
-import { ApiUsuariosService } from './api-usuarios.service';
-
-describe('ApiUsuariosService', () => {
-  let service: ApiUsuariosService;
-
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-    service = TestBed.inject(ApiUsuariosService);
-  });
-
-  it('should be created', () => {
-    expect(service).toBeTruthy();
-  });
-});
